Add files via upload
Figure 1. Schematic representation of two scenarios of host-switching. In scenario a. phylogeny with the species pool occurring on a regional spatial scale tends to have the largest similarity between rodent species, but the opportunity of infection is low since sister rodent species do not necessarily co-occur. In scenario b. species phylogenies from each local empirical study have less similarity between species when compared to the phylogeny of the regional pool, but the opportunity for interaction is higher within a community.